Move RunResult to the interfaces.
