+++ src/bmi/api.py
@@ -3,13 +3,27 @@ import bmi.benchmark.api as benchmark
 import bmi.estimators.api as estimators
 import bmi.samplers.api as samplers
 import bmi.transforms.api as transforms
-from bmi.interface import IMutualInformationPointEstimator, ISampler
+
+# ISort doesn't want to split these into several lines, conflicting with Black
+# isort: off
+from bmi.interface import (
+    IMutualInformationPointEstimator,
+    ISampler,
+    ITaskEstimator,
+    Pathlike,
+    RunResult,
+)
+
+# isort: on
 
 __all__ = [
     "benchmark",
     "estimators",
     "samplers",
     "transforms",
-    "ISampler",
     "IMutualInformationPointEstimator",
+    "ISampler",
+    "ITaskEstimator",
+    "Pathlike",
+    "RunResult",
 ]
