+++ src/bmi/benchmark/api.py
@@ -1,4 +1,4 @@
-from bmi.benchmark.core import RunResult, Task, TaskMetadata, generate_task
+from bmi.benchmark.core import Task, TaskMetadata, generate_task
 from bmi.benchmark.tasks.api import BENCHMARK_TASKS, save_benchmark_tasks
 from bmi.benchmark.tasks.spiral import generate_spiral_invariance_task
 
@@ -23,6 +23,5 @@ __all__ = [
     "REstimatorLNN",
     "Task",
     "TaskMetadata",
-    "RunResult",
     "BENCHMARK_TASKS",
 ]
